requirements.txt
dxfgrabber
python-dotenv
ezdxf
matplotlib
shapely
pyautocad